style: Simplify mock logic in UTXO RPC relayer
Ignore the "MOCK" environment flag. Use the cache only if the node
does not know the UTXO and the "mock" feature flag is set. It is safe
to rely on node's response when the feature flag is not set because
the only case in which the cache contains an entry that the node does
not know about is when it was imagined previously, which can only
happen if mocking is enabled.