fix(RPC): Encode hashmap as list for transmission
In commit 1096a293 on neptune-core, the return type of RPC endpoint
`addition_records_for_block` changes from `HashMap` to `Vec`. This
commit applies the matching change on the side of the client,
including building the original hashmap from the transmitted list.

These commits (the present one and the one referenced) fix the issue
that transparent transaction info could not be transferred across the
RPC layer.