feat: Add rate-limiting as countermeasure against flooding
Add rate-limiting logic to give every IP a quotum of 1 request per 10
milliseconds. The IP is obtained from the socket address.

When the feature flag "attacks" is enabled on the server, and the HTTP header
field "X-Real-IP-Override" on the request is set, then the server uses the field
value instead of obtaining it from the socket address. This configuration allows
the rate-limiting logic to be tested.

To test without rate-limiting:
 1. Open 4 terminal windows.
 2. Start neptune-core in one.
 3. Start neptune-explorer in two: `cargo run -- --site-domain test`
 4. Start the scraper in three: `cargo run --bin scraper --features="attacks"`
    and observe the green info logs indicating successful retrieval of web
    pages.
 5. Start the attack in four:
    `cargo run --bin high_rate_attack --features="attacks"`.
    Observe that in window three the green info messages turn into orange warn
    messages, indicating response timeouts. Also, try opening an explorer page.
    It gives a timeout.

To test with rate-limiting:
 3. Start neptune-explorer in two, but with feature flag "attacks" enabled:
    `cargo run --features="attacks" -- --site-domain test`
 5. Start the attack and observe that the green info messages continue without
    interruption. Opening an explorer page does not result in a timeout.